requirements.txt
streamlit>=1.28.0
python-pptx>=0.6.21
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.18.0
streamlit-modal>=0.1.0
